pyproject.toml
[tool.poetry]
name = "zaberxygui"
version = "0.1.0"
description = ""
authors = ["Piet Swinkels <org@swnkls.nl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pysimplegui = "^4.60.5"
matplotlib = "^3.8.0"
numpy = "^1.26.0"
zaber-motion = "^4.4.0"
pyserial = "^3.5"
ipykernel = "^6.25.2"
pandas = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
